Feature/transactions merge - implemented authorizing codes, fully tested

+++ src/main/java/pl/zzpwj_2022_mu_pc_pk_sr/websitebank2022/repository/BankAccountTypeRepository.java
@@ -10,5 +10,5 @@ import java.util.Optional;
 
 @Repository
 public interface BankAccountTypeRepository extends JpaRepository<BankAccountType, Long> {
-    Optional<BankAccountType> findById(int id);
+    Optional<BankAccountType> findById(Long id);
 }
